# panel-api/internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: Jabali Automation API
  version: "v1"
  description: |
    The **stable, documented subset** of the Jabali Panel REST API intended for
    programmatic / token automation (scripts, CI, routers). It is **not** an
    exhaustive reference for every backend endpoint the web UI calls — the panel
    has many more internal, session-only endpoints (server status, security,
    backups, notifications, apps, …) intentionally outside this contract.

    Two authentication paths:

    - **Browser sessions** (Kratos cookie) — interactive UI use
    - **Bearer tokens** (`Authorization: Bearer jat_…`) — programmatic access
      from scripts / CI / routers. Tokens are scoped to the user that minted
      them; admin tokens carry the same is_admin flag as the owning user.

    Every endpoint enforces ownership (`claims.UserID == resource.UserID`)
    OR `is_admin = true`. Endpoints under `/admin/` are admin-only.

    Token mint UI: tenant shell → API Tokens. Plaintext is shown once at
    creation; the panel only stores the SHA-256 hash thereafter.

    ## Error envelope

    All non-2xx responses return JSON:
    ```
    { "error": "<machine-readable>", "message": "<human-readable, optional>" }
    ```
    HTTP status conveys the category; the `error` field disambiguates within
    a category (e.g. 401 may be `missing_credentials` vs `invalid_token`).
  contact:
    name: Jabali Panel
    url: https://git.jabali-panel.com/shukivaknin/jabali2
servers:
...
components:
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: jat_…
      description: |
        Per-user token minted via POST /me/api-tokens. Send as
        `Authorization: Bearer jat_<43-char base64url>`.
    KratosSession:
      type: apiKey
      in: cookie
      name: ory_kratos_session
      description: Browser cookie from /login. Same-origin only.
...
security:
  - BearerToken: []
  - KratosSession: []

# -------------------------------------------------------------------- paths ---
paths:
...
  /nic/update:
    get:
      tags: [DDNS]
      summary: DynDNS v2 — update an A/AAAA record
      description: |
        Speaks the de-facto DynDNS protocol every router firmware
        (OpenWrt, Mikrotik, pfSense, Ubiquiti, ddclient, inadyn) already
        knows. Wraps the same per-user Bearer token in HTTP Basic-auth
        so existing DDNS clients can update records without a custom
        scripting integration.

        **Path:** mounted at the engine root, NOT under /api/v1, because
        every router hardcodes /nic/update.

        **Auth:** HTTP Basic. The username half is ignored (routers
        require it as input but the panel uses only the password). The
        password is your `jat_…` token.

        **myip:** omit to use the client's source IP (X-Real-IP /
        X-Forwarded-For / RemoteAddr).

        **Hostname:** must be a fully-qualified domain name. The panel
        does longest-suffix match against your owned domains to find
        the zone + record.
      servers:
        - url: https://{panel-hostname}:8443
          variables: { panel-hostname: { default: jabali-panel.local } }
      security:
        - BasicAuth: []
      parameters:
        - in: query
          name: hostname
          required: true
          schema: { type: string, example: vpn.example.com }
          description: FQDN of the record to update.
        - in: query
          name: myip
          schema: { type: string, example: 203.0.113.5 }
          description: New IP. Omit to use the request's source IP.
      responses:
        "200":
          description: |
            Text/plain single-line response — one of:
            - `good 1.2.3.4` — record updated
            - `nochg 1.2.3.4` — record already pointed at that IP
          content:
            text/plain:
              schema: { type: string, example: "good 1.2.3.4\n" }
        "400":
          description: "Malformed hostname or IP. Body: `notfqdn`."
          content: { text/plain: { schema: { type: string, example: "notfqdn\n" } } }
        "401":
          description: "Missing or invalid Basic auth. Body: `badauth`."
          content: { text/plain: { schema: { type: string, example: "badauth\n" } } }
        "404":
          description: "Hostname doesn't resolve to a record you own. Body: `nohost`."
          content: { text/plain: { schema: { type: string, example: "nohost\n" } } }
    post:
      tags: [DDNS]
      summary: DynDNS v2 — same as GET (some routers POST)
      description: Some router firmware sends POST instead of GET; semantics identical.
      security: [ { BasicAuth: [] } ]
      responses:
        "200": { description: See GET }

  # ------------------------------------------------------------- Domains ---
...